projects/modules/04-fastapi-web/requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.27
pydantic>=2.6
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
sqlalchemy>=2.0
httpx>=0.27
pytest>=8.0
